Add `Encoding::{from_bytes, to_bytes}`
Adds provided methods to the `Encoding` trait which take a `ByteOrder`
value to determine the order.

This also factors more encoding-related functionality into the new
`encoding` module.
